# openapi.yaml
openapi: 3.0.3
info:
  title: 医养结合云端后台管理系统 API
  version: 2.0.0
  description: |
    基于 train.md 与当前 Spring Boot 实现维护的 REST API 契约。
    除登录、公开注册和刷新令牌外，接口均需要 Bearer access token。
    JSON 成功响应的业务 code 固定为 0，错误响应同时使用正确的 HTTP 状态码。
servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
security:
  - bearerAuth: []
...
paths:
...
  /key-populations:
    get:
      tags: [KeyPopulation]
      summary: 重点人群列表
      operationId: listKeyPopulations
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [active, closed] } }
        - { $ref: '#/components/parameters/PageNo' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiKeyPopulationPage' }
    post:
      tags: [KeyPopulation]
      summary: 新增重点人群
      operationId: createKeyPopulation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/KeyPopulationCreateRequest' }
      responses:
        '201': { $ref: '#/components/responses/KeyPopulationResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /key-populations/{id}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [KeyPopulation]
      summary: 重点人群详情
      operationId: getKeyPopulation
      responses:
        '200': { $ref: '#/components/responses/KeyPopulationResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [KeyPopulation]
      summary: 更新重点人群
      operationId: updateKeyPopulation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/KeyPopulationUpdateRequest' }
      responses:
        '200': { $ref: '#/components/responses/KeyPopulationResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [KeyPopulation]
      summary: 软删除重点人群记录
      operationId: deleteKeyPopulation
      responses:
        '200': { $ref: '#/components/responses/EmptyResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /key-populations/{id}/close:
    post:
      tags: [KeyPopulation]
      summary: 关闭重点人群记录
      operationId: closeKeyPopulation
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/EmptyResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /users:
    get:
      tags: [UserAccount]
      summary: 系统账户列表
      operationId: listUsers
      parameters:
        - { $ref: '#/components/parameters/Keyword' }
        - { name: roleCode, in: query, schema: { $ref: '#/components/schemas/RoleCode' } }
        - { name: status, in: query, schema: { type: string, enum: [enabled, disabled] } }
        - { $ref: '#/components/parameters/PageNo' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiUserPage' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [UserAccount]
      summary: 创建系统账户
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserCreateRequest' }
      responses:
        '201': { $ref: '#/components/responses/UserResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /users/{id}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [UserAccount]
      summary: 系统账户详情
      operationId: getUser
      responses:
        '200': { $ref: '#/components/responses/UserResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [UserAccount]
      summary: 更新系统账户
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdateRequest' }
      responses:
        '200': { $ref: '#/components/responses/UserResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [UserAccount]
      summary: 软删除系统账户
      operationId: deleteUser
      responses:
        '200': { $ref: '#/components/responses/EmptyResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /users/{id}/status:
    post:
      tags: [UserAccount]
      summary: 修改系统账户状态
      operationId: updateUserStatus
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        $ref: '#/components/requestBodies/UserStatus'
      responses:
        '200': { $ref: '#/components/responses/EmptyResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /users/{id}/roles:
    post:
      tags: [UserAccount]
      summary: 分配单一系统角色
      operationId: assignUserRole
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoleAssignRequest' }
      responses:
        '200': { $ref: '#/components/responses/EmptyResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /profile:
    get:
      tags: [Profile]
      summary: 获取个人资料
      operationId: getProfile
      responses:
        '200': { $ref: '#/components/responses/UserResponse' }
    put:
      tags: [Profile]
      summary: 更新个人资料
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileUpdateRequest' }
      responses:
        '200': { $ref: '#/components/responses/UserResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /profile/change-password:
    post:
      tags: [Profile]
      summary: 修改当前用户密码
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePasswordRequest' }
      responses:
        '200': { $ref: '#/components/responses/EmptyResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /profile/elderly:
    get:
      tags: [Profile]
      summary: 当前老人账号的关联档案
      operationId: getMyElderlyProfile
      responses:
        '200': { $ref: '#/components/responses/ElderlyResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, maxLength: 64 }
    Keyword:
      name: keyword
      in: query
      schema: { type: string, maxLength: 100 }
    PageNo:
      name: pageNo
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
  requestBodies:
...
    UserStatus:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/UserStatusUpdateRequest' }
  responses:
    EmptyResponse:
      description: 成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiEmpty' }
...
    UserResponse:
      description: 成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiUser' }
    ElderlyResponse:
      description: 成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiElderly' }
...
    KeyPopulationResponse:
      description: 成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiKeyPopulation' }
...
    BadRequest:
      description: 参数或业务输入错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
    Forbidden:
      description: 当前角色无权访问
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: 数据不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Conflict:
      description: 数据冲突或仍被引用
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
    ApiBase:
      type: object
      required: [code, message, traceId]
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: success }
        traceId: { type: string, example: b3c1f8e9d8f14a7f }
    ApiEmpty:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { nullable: true }
    ApiError:
      type: object
      required: [code, message, traceId]
      properties:
        code: { type: integer, enum: [400, 401, 403, 404, 409, 500, 501] }
        message: { type: string }
        data: { nullable: true }
        traceId: { type: string }
...
    RoleCode:
      type: string
      enum: [admin, doctor, operator, analyst, elderly]
    User:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        realName: { type: string }
        roleCode: { $ref: '#/components/schemas/RoleCode' }
        mobile: { type: string, nullable: true }
        institutionId: { type: string, nullable: true }
        regionCode: { type: string, nullable: true }
        status: { type: string, enum: [enabled, disabled] }
        lastLoginAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    UserCreateRequest:
      type: object
      required: [username, password, realName, roleCode]
      properties:
        username: { type: string, minLength: 3, maxLength: 50, pattern: '^[A-Za-z0-9_]+$' }
        password: { type: string, minLength: 8, maxLength: 64, writeOnly: true }
        realName: { type: string, maxLength: 50 }
        roleCode: { $ref: '#/components/schemas/RoleCode' }
        mobile: { type: string, pattern: '^1[3-9]\d{9}$' }
        institutionId: { type: string, maxLength: 64 }
        regionCode: { type: string, maxLength: 64 }
        status: { type: string, enum: [enabled, disabled], default: enabled }
...
    UserUpdateRequest:
      type: object
      properties:
        realName: { type: string, maxLength: 50 }
        roleCode: { $ref: '#/components/schemas/RoleCode' }
        mobile: { type: string, pattern: '^1[3-9]\d{9}$' }
        institutionId: { type: string, maxLength: 64 }
        regionCode: { type: string, maxLength: 64 }
        status: { type: string, enum: [enabled, disabled] }
    UserStatusUpdateRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [enabled, disabled] }
    RoleAssignRequest:
      type: object
      required: [roleCode]
      properties:
        roleCode: { $ref: '#/components/schemas/RoleCode' }
...
    ChangePasswordRequest:
      type: object
      required: [oldPassword, newPassword]
      properties:
        oldPassword: { type: string, writeOnly: true }
        newPassword: { type: string, minLength: 8, maxLength: 64, writeOnly: true }
    ProfileUpdateRequest:
      type: object
      properties:
        realName: { type: string, maxLength: 50 }
        mobile: { type: string, pattern: '^1[3-9]\d{9}$' }
    ApiUser:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { $ref: '#/components/schemas/User' }
    UserPage:
      type: object
      required: [list, pageNo, pageSize, total]
      properties:
        list: { type: array, items: { $ref: '#/components/schemas/User' } }
        pageNo: { type: integer }
        pageSize: { type: integer }
        total: { type: integer, format: int64 }
    ApiUserPage:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { $ref: '#/components/schemas/UserPage' }
    ElderlyProfile:
      type: object
      properties:
        id: { type: string }
        userId: { type: string, nullable: true }
        name: { type: string }
        gender: { type: string, enum: [male, female, unknown] }
        birthday: { type: string, format: date, nullable: true }
        age: { type: integer, nullable: true }
        phone: { type: string, nullable: true }
        address: { type: string, nullable: true }
        institutionId: { type: string, nullable: true }
        regionCode: { type: string, nullable: true }
        medicalHistory: { type: string, nullable: true }
        careLevel: { type: string, nullable: true }
        tags: { type: array, items: { type: string } }
        status: { type: string, enum: [active, inactive] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ApiElderly:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { $ref: '#/components/schemas/ElderlyProfile' }
...
    KeyPopulation:
      type: object
      properties:
        id: { type: string }
        elderlyId: { type: string }
        category: { type: string }
        reason: { type: string, nullable: true }
        level: { type: string }
        ownerDoctorId: { type: string, nullable: true }
        followUpCycleDays: { type: integer }
        status: { type: string, enum: [active, closed] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    KeyPopulationCreateRequest:
      type: object
      required: [elderlyId, category]
      properties:
        elderlyId: { type: string, maxLength: 64 }
        category: { type: string, maxLength: 50 }
        reason: { type: string, maxLength: 500 }
        level: { type: string, maxLength: 10, default: B }
        ownerDoctorId: { type: string, maxLength: 64 }
        followUpCycleDays: { type: integer, default: 30 }
        status: { type: string, enum: [active, closed], default: active }
    KeyPopulationUpdateRequest:
      type: object
      properties:
        category: { type: string, maxLength: 50 }
        reason: { type: string, maxLength: 500 }
        level: { type: string, maxLength: 10 }
        ownerDoctorId: { type: string, maxLength: 64 }
        followUpCycleDays: { type: integer }
        status: { type: string, enum: [active, closed] }
    ApiKeyPopulation:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { $ref: '#/components/schemas/KeyPopulation' }
...
    KeyPopulationPage:
      type: object
      properties:
        list: { type: array, items: { $ref: '#/components/schemas/KeyPopulation' } }
        pageNo: { type: integer }
        pageSize: { type: integer }
        total: { type: integer, format: int64 }
    ApiKeyPopulationPage:
      allOf:
        - { $ref: '#/components/schemas/ApiBase' }
        - type: object
          properties:
            data: { $ref: '#/components/schemas/KeyPopulationPage' }
...